Fixed ordering of phase subblocks
The order of the phases in the input file is meant
to determine the identity of the first and second
phase. Before the fix, the list of phase sub-blocks
was created using the keys of a dictionary, but these
are unordered, so this was not correct. A separate
list now captures the order in the input file.